rename DestinationAcceptanceTest into AbstractDestinationAcceptanceTest
